fix(notional-finance-v3): Remove fCash app token and move logic into supply and borrow positions
