+++ src/apps/notional-finance-v3/arbitrum/notional-finance-v3.supply.contract-position-fetcher.ts
@@ -32,9 +32,13 @@ export type NotionalFinanceLendingDataProps = {
   currencyId: number;
   tokenId: string;
   maturity: number;
+  fCashPV: number;
   positionKey: string;
   type: string;
 };
+
+const NOTIONAL_CONSTANT = 10 ** 7;
+const SECONDS_IN_YEAR = 31104000; // 360 days
 
 @PositionTemplate()
 export class ArbitrumNotionalFinanceV3SupplyContractPositionFetcher extends ContractPositionTemplatePositionFetcher<
@@ -81,11 +85,19 @@ export class ArbitrumNotionalFinanceV3SupplyContractPositionFetcher extends Cont
               .read.encodeToId([currencyId, maturity, 0])
               .then(v => v.toString());
 
+            const apy = Number(activeMarket.lastImpliedRate ?? 0) / NOTIONAL_CONSTANT;
+
+            const dateNowEpoch = Date.now() / 1000;
+            const timeToMaturity = (maturity - dateNowEpoch) / SECONDS_IN_YEAR;
+            const lastImpliedRate = apy / 100;
+            const fCashPV = 1 / Math.exp(lastImpliedRate * timeToMaturity);
+
             return {
               address: this.notionalViewContractAddress,
               currencyId,
               underlyingTokenAddress,
               maturity,
+              fCashPV,
               tokenId,
               type,
             };
@@ -100,16 +112,14 @@ export class ArbitrumNotionalFinanceV3SupplyContractPositionFetcher extends Cont
   }
 
   async getTokenDefinitions({ definition }: GetTokenDefinitionsParams<NotionalView, NotionalFinanceLendingDefinition>) {
-    return [
-      { metaType: MetaType.SUPPLIED, address: definition.address, network: this.network, tokenId: definition.tokenId },
-    ];
+    return [{ metaType: MetaType.SUPPLIED, address: definition.underlyingTokenAddress, network: this.network }];
   }
 
   async getDataProps(
     params: GetDataPropsParams<NotionalView, NotionalFinanceLendingDataProps, NotionalFinanceLendingDefinition>,
   ) {
     const defaultDataProps = await super.getDataProps(params);
-    const props = pick(params.definition, ['currencyId', 'tokenId', 'maturity', 'type']);
+    const props = pick(params.definition, ['currencyId', 'tokenId', 'maturity', 'fCashPV', 'type']);
     return { ...defaultDataProps, ...props, positionKey: Object.values(props).join(':') };
   }
 
@@ -122,7 +132,7 @@ export class ArbitrumNotionalFinanceV3SupplyContractPositionFetcher extends Cont
     contractPosition,
     contract,
   }: GetTokenBalancesParams<NotionalView, NotionalFinanceLendingDataProps>): Promise<BigNumberish[]> {
-    const { maturity, currencyId } = contractPosition.dataProps;
+    const { maturity, currencyId, fCashPV } = contractPosition.dataProps;
     const portfolio = await contract.read.getAccountPortfolio([address]);
     const supplyPositions = portfolio.filter(v => v.notional >= 0);
     const position = supplyPositions.find(v => Number(v.maturity) === maturity && Number(v.currencyId) === currencyId);
@@ -132,6 +142,8 @@ export class ArbitrumNotionalFinanceV3SupplyContractPositionFetcher extends Cont
       .times(10 ** contractPosition.tokens[0].decimals)
       .div(10 ** 8);
 
-    return [fcashAmountAtMaturity.abs().toString()];
+    const presentValue = fcashAmountAtMaturity.times(new BigNumber(fCashPV));
+
+    return [presentValue.toString()];
   }
 }
