+++ src/apps/notional-finance-v3/notional-finance-v3.module.ts
@@ -3,7 +3,6 @@ import { Module } from '@nestjs/common';
 import { AbstractApp } from '~app/app.dynamic-module';
 
 import { ArbitrumNotionalFinanceV3BorrowContractPositionFetcher } from './arbitrum/notional-finance-v3.borrow.contract-position-fetcher';
-import { ArbitrumNotionalFinanceV3FCashTokenFetcher } from './arbitrum/notional-finance-v3.f-cash.token-fetcher';
 import { ArbitrumNotionalFinanceV3NTokenTokenFetcher } from './arbitrum/notional-finance-v3.n-token.token-fetcher';
 import { ArbitrumNotionalFinanceV3PCashTokenFetcher } from './arbitrum/notional-finance-v3.p-cash.token-fetcher';
 import { ArbitrumNotionalFinanceV3PDebtTokenFetcher } from './arbitrum/notional-finance-v3.p-debt.token-fetcher';
@@ -17,7 +16,6 @@ import { NotionalFinanceV3ViemContractFactory } from './contracts';
     ArbitrumNotionalFinanceV3PCashTokenFetcher,
     ArbitrumNotionalFinanceV3PDebtTokenFetcher,
     ArbitrumNotionalFinanceV3NTokenTokenFetcher,
-    ArbitrumNotionalFinanceV3FCashTokenFetcher,
     ArbitrumNotionalFinanceV3SupplyContractPositionFetcher,
     ArbitrumNotionalFinanceV3BorrowContractPositionFetcher,
   ],
